requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0

# Tests et qualité de code
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

